# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResBus": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"BussName": {
						"properties": {
							"type": {
								"type": "string"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					},
					"BussSeats": {
						"properties": {
							"type": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					},
					"BussBookingDates": {
						"properties": {
							"End_date": {
								"properties": {
									"type": {
										"type": "string"
									}
								},
								"required": [
									"type"
								],
								"type": "object"
							},
							"Start_date": {
								"properties": {
									"type": {
										"type": "string"
									}
								},
								"required": [
									"type"
								],
								"type": "object"
							}
						},
						"required": [
							"End_date",
							"Start_date"
						],
						"type": "object"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"BussName",
					"BussSeats",
					"BussBookingDates",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveBus": {
				"properties": {
					"BussName": {
						"type": "string"
					},
					"BussSeats": {
						"type": "number",
						"format": "double"
					},
					"BussBookingDates": {
						"properties": {
							"End_date": {
								"type": "string"
							},
							"Start_date": {
								"type": "string"
							}
						},
						"required": [
							"End_date",
							"Start_date"
						],
						"type": "object"
					}
				},
				"required": [
					"BussName",
					"BussSeats",
					"BussBookingDates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetBus": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBus": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"BussName": {
						"type": "string"
					},
					"BussSeats": {
						"type": "number",
						"format": "double"
					},
					"BussBookingDates": {
						"properties": {
							"End_date": {
								"type": "string"
							},
							"Start_date": {
								"type": "string"
							}
						},
						"required": [
							"End_date",
							"Start_date"
						],
						"type": "object"
					}
				},
				"required": [
					"_id",
					"BussName",
					"BussSeats",
					"BussBookingDates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteBus": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveTour": {
				"properties": {
					"no_of_participants": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"no_of_participants"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "http",
				"name": "Authorization",
				"in": "header",
				"scheme": "Bearer"
			}
		}
	},
	"info": {
		"title": "webapi",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/Bus/savenewbus": {
			"post": {
				"operationId": "Savebus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBus"
								}
							}
						}
					}
				},
				"tags": [
					"Bus"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveBus"
							}
						}
					}
				}
			}
		},
		"/Bus/getbus": {
			"post": {
				"operationId": "Getbus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBus"
								}
							}
						}
					}
				},
				"tags": [
					"Bus"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetBus"
							}
						}
					}
				}
			}
		},
		"/Bus/updatebus": {
			"put": {
				"operationId": "Updatebus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBus"
								}
							}
						}
					}
				},
				"tags": [
					"Bus"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBus"
							}
						}
					}
				}
			}
		},
		"/Bus/deletebus": {
			"delete": {
				"operationId": "Deletebus",
				"responses": {
					"200": {
						"description": "Product deleted",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Bus"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteBus"
							}
						}
					}
				}
			}
		},
		"/Bus/getbuslist": {
			"post": {
				"operationId": "GetbusList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResBus"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bus"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/Tour/savetourinfo": {
			"post": {
				"operationId": "Savetour",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Tour endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveTour"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}